udc_backend_flask/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.7.0
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-MQTT==1.2.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
greenlet==3.1.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.0.0
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.36
statsmodels==0.14.4
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
